+++ actions/filter-files/package.json
@@ -1,7 +1,6 @@
 {
     "name": "filter-files",
     "version": "0.3.0",
-    "type": "module",
     "dependencies": {
         "picomatch": "2"
     }
