+++ package.json
@@ -10,7 +10,6 @@
         "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --testRegex '.*.test.mjs'"
     },
     "devDependencies": {
-        "@babel/preset-env": "^7.22.20",
         "@changesets/cli": "^2.22.0",
         "@khanacademy/eslint-config": "^0.1.0",
         "@vercel/ncc": "^0.36.1",
